// client/src/storehouse/Action.ts
export enum Action {
    CALL = 'CALL',
    ADD_CONTACT = 'ADD_CONTACT',
    EDIT_CONTACT = 'EDIT_CONTACT',
    SET_CONTACTS = 'SET_CONTACTS',
    FETCH_CONTACTS = 'FETCH_CONTACTS',
    REMOVE_CONTACT = 'REMOVE_CONTACT',
    SET_CALLS_HISTORY = 'SET_CALLS_HISTORY',
    TOGGLE_REG_WINDOW = 'TOGGLE_REG_WINDOW',
    TOGGLE_LOGIN_WINDOW = 'TOGGLE_LOGIN_WINDOW',
    TOGGLE_CONTACT_WINDOW = 'TOGGLE_CONTACT_WINDOW',
    TOGGLE_ADD_CONTACT_WINDOW = 'TOGGLE_ADD_CONTACT_WINDOW',
    TOGGLE_EDIT_CONTACT_WINDOW = 'TOGGLE_EDIT_CONTACT_WINDOW'
}

export default Action;